pyproject.toml
[project]
name = "ai-recruitment-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "fastapi==0.129.0",
    "frozenlist==1.8.0",
    "greenlet==3.3.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.11",
    "langchain-openai==1.1.9",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.8",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.7",
    "langgraph-sdk==0.3.5",
    "langsmith==0.7.1",
    "marshmallow==3.26.2",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.2",
    "openai==2.20.0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pdf2image==1.17.0",
    "pillow==12.1.1",
    "propcache==0.4.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pypdf>=6.7.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "regex==2026.1.15",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "starlette==0.52.1",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "uvicorn==0.40.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
